refactor: Improve last commit
