+++ packages/ui/src/components/SideMenu/index.ts
@@ -1,1 +1,2 @@
 export * from './SideMenu'
+export * from './ToggleMenuButton'
